requirements.txt
boto3
loguru
pydantic
pyyaml
smart_open
pandas>=1.2.2